refactor(configs): rewrite Server class to a createServer function
